requirements.txt
Pillow==8.1.0
gpxpy==1.4.2
SRTM.py==0.3.6
